+++ src/icons/all-filled.js
@@ -38,6 +38,8 @@ import cloudFill from "./filled/cloud.svg";
 import codeFill from "./filled/code.svg";
 import compassFill from "./filled/compass.svg";
 import copyFill from "./filled/copy.svg";
+import currencyFill from "./filled/currency.svg";
+import cropFill from "./filled/crop.svg";
 import discordFill from "./filled/discord.svg";
 import downFromBracketFill from "./filled/down-from-bracket.svg";
 import downToBracketFill from "./filled/down-to-bracket.svg";
@@ -55,6 +57,11 @@ import fanFill from "./filled/fan.svg";
 import fastBackFill from "./filled/fast-back.svg";
 import fastForwardFill from "./filled/fast-forward.svg";
 import figmaFill from "./filled/figma.svg";
+import fileFill from "./filled/file.svg";
+import filePlusFill from "./filled/file-plus.svg";
+import fileTextFill from "./filled/file-text.svg";
+import flipHorizontalFill from "./filled/flip-horizontal.svg";
+import flipVerticalFill from "./filled/flip-vertical.svg";
 import flaskFill from "./filled/flask.svg";
 import floppyDiskFill from "./filled/floppy-disk.svg";
 import flowerFill from "./filled/flower.svg";
@@ -91,6 +98,8 @@ import plugFill from "./filled/plug.svg";
 import plusFill from "./filled/plus.svg";
 import puzzleFill from "./filled/puzzle.svg";
 import redoFill from "./filled/redo.svg";
+import rotateLeftFill from "./filled/rotate-left.svg";
+import rotateRightFill from "./filled/rotate-right.svg";
 import rightFromBracketFill from "./filled/right-from-bracket.svg";
 import rightToBracketFill from "./filled/right-to-bracket.svg";
 import robotFill from "./filled/robot.svg";
@@ -122,6 +131,7 @@ import upFromBracketFill from "./filled/up-from-bracket.svg";
 import upToBracketFill from "./filled/up-to-bracket.svg";
 import userFill from "./filled/user.svg";
 import usersFill from "./filled/users.svg";
+import walletFill from "./filled/wallet.svg";
 import waveformFill from "./filled/waveform.svg";
 import wrenchFill from "./filled/wrench.svg";
 import xFill from "./filled/x.svg";
@@ -156,6 +166,8 @@ registerIcons({
     "code-fill": codeFill,
     "compass-fill": compassFill,
     "copy-fill": copyFill,
+    "currency-fill": currencyFill,
+    "crop-fill": cropFill,
     "discord-fill": discordFill,
     "down-from-bracket-fill": downFromBracketFill,
     "down-to-bracket-fill": downToBracketFill,
@@ -173,6 +185,11 @@ registerIcons({
     "fast-back-fill": fastBackFill,
     "fast-forward-fill": fastForwardFill,
     "figma-fill": figmaFill,
+    "file-fill": fileFill,
+    "file-plus-fill": filePlusFill,
+    "file-text-fill": fileTextFill,
+    "flip-horizontal-fill": flipHorizontalFill,
+    "flip-vertical-fill": flipVerticalFill,
     "flask-fill": flaskFill,
     "floppy-disk-fill": floppyDiskFill,
     "flower-fill": flowerFill,
@@ -209,6 +226,8 @@ registerIcons({
     "plus-fill": plusFill,
     "puzzle-fill": puzzleFill,
     "redo-fill": redoFill,
+    "rotate-left-fill": rotateLeftFill,
+    "rotate-right-fill": rotateRightFill,
     "right-from-bracket-fill": rightFromBracketFill,
     "right-to-bracket-fill": rightToBracketFill,
     "robot-fill": robotFill,
@@ -240,6 +259,7 @@ registerIcons({
     "up-to-bracket-fill": upToBracketFill,
     "user-fill": userFill,
     "users-fill": usersFill,
+    "wallet-fill": walletFill,
     "waveform-fill": waveformFill,
     "wrench-fill": wrenchFill,
     "x-fill": xFill,